+++ dymension/libs/kaspa/lib/validator/src/lib.rs
@@ -2,3 +2,5 @@ pub mod confirmation;
 pub mod deposit;
 pub mod withdraw;
 pub mod signer;
+
+pub use secp256k1::Keypair as KaspaSecpKeypair;
